Rm separate file
